Remove custom software type and use kotlinApplication with features
