+++ build-logic/plugins/src/main/java/org/gradle/client/ecosystem/CustomEcosystemPlugin.java
@@ -3,14 +3,12 @@ package org.gradle.client.ecosystem;
 import org.gradle.api.Plugin;
 import org.gradle.api.initialization.Settings;
 import org.gradle.api.internal.plugins.software.RegistersSoftwareTypes;
-import org.gradle.client.softwaretype.CustomDesktopComposeApplicationPlugin;
-import org.gradle.client.softwaretype.compose.ComposeSoftwareFeaturePlugin;
-import org.gradle.client.softwaretype.detekt.DetektSoftwareFeaturePlugin;
-import org.gradle.client.softwaretype.sqldelight.SqlDelightSoftwareFeaturePlugin;
+import org.gradle.client.softwarefeatures.compose.ComposeSoftwareFeaturePlugin;
+import org.gradle.client.softwarefeatures.detekt.DetektSoftwareFeaturePlugin;
+import org.gradle.client.softwarefeatures.sqldelight.SqlDelightSoftwareFeaturePlugin;
 
 @SuppressWarnings("UnstableApiUsage")
 @RegistersSoftwareTypes({
-        CustomDesktopComposeApplicationPlugin.class,
         ComposeSoftwareFeaturePlugin.class,
         DetektSoftwareFeaturePlugin.class,
         SqlDelightSoftwareFeaturePlugin.class
