+++ build-logic/plugins/src/main/java/org/gradle/client/softwarefeatures/compose/ComposeBuildModel.java
@@ -1,0 +1,10 @@
+package org.gradle.client.softwarefeatures.compose;
+
+import org.gradle.api.internal.plugins.BuildModel;
+import org.jetbrains.compose.ComposeExtension;
+
+public interface ComposeBuildModel extends BuildModel {
+
+    ComposeExtension getComposeExtension();
+
+}
