+++ build-logic/plugins/src/main/java/org/gradle/client/softwarefeatures/compose/MacOS.java
@@ -1,22 +1,19 @@
-package org.gradle.client.softwaretype.compose;
+package org.gradle.client.softwarefeatures.compose;
 
 import org.gradle.api.file.RegularFileProperty;
 import org.gradle.api.provider.Property;
 import org.gradle.declarative.dsl.model.annotations.Restricted;
 
-public interface Windows {
+public interface MacOS {
     @Restricted
     RegularFileProperty getIconFile();
 
     @Restricted
-    Property<Boolean> getMenu();
+    Property<Boolean> getAppStore();
 
     @Restricted
-    Property<String> getMenuGroup();
+    Property<String> getBundleID();
 
     @Restricted
-    Property<Boolean> getPerUserInstall();
-
-    @Restricted
-    Property<String> getUpgradeUuid();
+    Property<String> getDockName();
 }
