+++ build-logic/plugins/src/main/java/org/gradle/client/softwarefeatures/compose/Module.java
@@ -1,4 +1,4 @@
-package org.gradle.client.softwaretype.compose;
+package org.gradle.client.softwarefeatures.compose;
 
 import org.gradle.api.Named;
 import org.gradle.api.provider.Property;
