+++ build-logic/plugins/src/main/java/org/gradle/client/softwarefeatures/detekt/DetektBuildModel.java
@@ -1,4 +1,4 @@
-package org.gradle.client.softwaretype.detekt;
+package org.gradle.client.softwarefeatures.detekt;
 
 import io.gitlab.arturbosch.detekt.extensions.DetektExtension;
 import org.gradle.api.internal.plugins.BuildModel;
