+++ build-logic/plugins/src/main/java/org/gradle/client/softwarefeatures/package-info.java
@@ -2,4 +2,4 @@
  * Contains types needed for a Desktop ComposeApplication Application {@link org.gradle.api.internal.plugins.software.SoftwareType SoftwareType}.
  */
 @org.jspecify.annotations.NullMarked
-package org.gradle.client.softwaretype;
+package org.gradle.client.softwarefeatures;
