+++ build-logic/plugins/src/main/java/org/gradle/client/softwarefeatures/sqldelight/SqlDelight.java
@@ -1,8 +1,7 @@
-package org.gradle.client.softwaretype.sqldelight;
+package org.gradle.client.softwarefeatures.sqldelight;
 
 import org.gradle.api.NamedDomainObjectContainer;
 import org.gradle.api.internal.plugins.HasBuildModel;
-import org.gradle.declarative.dsl.model.annotations.Restricted;
 import org.jspecify.annotations.NonNull;
 
 public interface SqlDelight extends HasBuildModel<@NonNull SqlDelightBuildModel> {
