+++ build-logic/plugins/src/main/java/org/gradle/client/softwarefeatures/sqldelight/SqlDelightBuildModel.java
@@ -1,11 +1,7 @@
-package org.gradle.client.softwaretype.sqldelight;
+package org.gradle.client.softwarefeatures.sqldelight;
 
 import app.cash.sqldelight.gradle.SqlDelightExtension;
-import org.gradle.api.NamedDomainObjectContainer;
 import org.gradle.api.internal.plugins.BuildModel;
-import org.gradle.api.provider.ProviderFactory;
-
-import javax.inject.Inject;
 
 public interface SqlDelightBuildModel extends BuildModel {
