+++ gradle-client/build.gradle.dcl
@@ -1,66 +1,64 @@
-desktopComposeApp {
+kotlinApplication {
     group = "org.gradle.client"
 
     // Version must be strictly x.y.z and >= 1.0.0
     // for native packaging to work across platforms
     version = "1.1.3"
 
-    kotlinApplication {
-        dependencies {
-            implementation(platform("org.jetbrains.kotlin:kotlin-bom:2.0.21"))
-            implementation(platform("org.jetbrains.kotlinx:kotlinx-coroutines-bom:1.8.1"))
-            implementation(platform("org.jetbrains.kotlinx:kotlinx-serialization-bom:1.6.3"))
-        }
+    dependencies {
+        implementation(platform("org.jetbrains.kotlin:kotlin-bom:2.0.21"))
+        implementation(platform("org.jetbrains.kotlinx:kotlinx-coroutines-bom:1.8.1"))
+        implementation(platform("org.jetbrains.kotlinx:kotlinx-serialization-bom:1.6.3"))
+    }
 
-        targets {
-            jvm {
-                jdkVersion = 17
+    targets {
+        jvm {
+            jdkVersion = 17
 
+            dependencies {
+                implementation(project(":build-action"))
+                implementation(project(":mutations-demo"))
+
+                implementation("org.gradle:gradle-tooling-api:8.12-20241009055624+0000")
+
+                implementation("com.arkivanov.decompose:decompose:3.0.0")
+                implementation("com.arkivanov.decompose:extensions-compose:3.0.0")
+                implementation("com.arkivanov.essenty:lifecycle-coroutines:1.3.0")
+                implementation("org.jetbrains.kotlinx:kotlinx-serialization-json:1.6.3")
+
+                implementation("io.ktor:ktor-client-okhttp:2.3.12")
+                implementation("io.ktor:ktor-serialization-kotlinx-json:2.3.12")
+
+                implementation("dev.chrisbanes.material3:material3-window-size-class-multiplatform:0.5.0")
+                implementation("com.materialkolor:material-kolor:1.7.0")
+                implementation("io.github.vinceglb:filekit-compose:0.8.2")
+
+                implementation("org.slf4j:slf4j-api:2.0.14")
+                implementation("ch.qos.logback:logback-classic:1.5.6")
+
+                implementation("org.gradle:gradle-declarative-dsl-core:9.0-milestone-1")
+                implementation("org.gradle:gradle-declarative-dsl-evaluator:9.0-milestone-1")
+                implementation("org.gradle:gradle-declarative-dsl-tooling-models:9.0-milestone-1")
+
+                runtimeOnly("org.jetbrains.kotlinx:kotlinx-coroutines-swing:1.8.1")
+
+                implementation("org.jetbrains.compose.runtime:runtime:1.6.11")
+                implementation("org.jetbrains.compose.foundation:foundation:1.6.11")
+                implementation("org.jetbrains.compose.material3:material3:1.6.11")
+                implementation("org.jetbrains.compose.material:material-icons-extended:1.6.11")
+                implementation("org.jetbrains.compose.ui:ui:1.6.11")
+                implementation("org.jetbrains.compose.components:components-resources:1.6.11")
+                implementation("org.jetbrains.compose.components:components-ui-tooling-preview:1.6.11")
+                implementation("org.jetbrains.compose.desktop:desktop-jvm-macos-arm64:1.6.11")
+                implementation("org.jetbrains.compose.desktop:desktop-jvm-linux-x64:1.6.11")
+                implementation("org.jetbrains.compose.desktop:desktop-jvm-windows-x64:1.6.11")
+            }
+
+            testing {
                 dependencies {
-                    implementation(project(":build-action"))
-                    implementation(project(":mutations-demo"))
+                    implementation("org.jetbrains.compose.ui:ui-test-junit4:1.6.11")
 
-                    implementation("org.gradle:gradle-tooling-api:8.12-20241009055624+0000")
-
-                    implementation("com.arkivanov.decompose:decompose:3.0.0")
-                    implementation("com.arkivanov.decompose:extensions-compose:3.0.0")
-                    implementation("com.arkivanov.essenty:lifecycle-coroutines:1.3.0")
-                    implementation("org.jetbrains.kotlinx:kotlinx-serialization-json:1.6.3")
-
-                    implementation("io.ktor:ktor-client-okhttp:2.3.12")
-                    implementation("io.ktor:ktor-serialization-kotlinx-json:2.3.12")
-
-                    implementation("dev.chrisbanes.material3:material3-window-size-class-multiplatform:0.5.0")
-                    implementation("com.materialkolor:material-kolor:1.7.0")
-                    implementation("io.github.vinceglb:filekit-compose:0.8.2")
-
-                    implementation("org.slf4j:slf4j-api:2.0.14")
-                    implementation("ch.qos.logback:logback-classic:1.5.6")
-
-                    implementation("org.gradle:gradle-declarative-dsl-core:9.0-milestone-1")
-                    implementation("org.gradle:gradle-declarative-dsl-evaluator:9.0-milestone-1")
-                    implementation("org.gradle:gradle-declarative-dsl-tooling-models:9.0-milestone-1")
-
-                    runtimeOnly("org.jetbrains.kotlinx:kotlinx-coroutines-swing:1.8.1")
-
-                    implementation("org.jetbrains.compose.runtime:runtime:1.6.11")
-                    implementation("org.jetbrains.compose.foundation:foundation:1.6.11")
-                    implementation("org.jetbrains.compose.material3:material3:1.6.11")
-                    implementation("org.jetbrains.compose.material:material-icons-extended:1.6.11")
-                    implementation("org.jetbrains.compose.ui:ui:1.6.11")
-                    implementation("org.jetbrains.compose.components:components-resources:1.6.11")
-                    implementation("org.jetbrains.compose.components:components-ui-tooling-preview:1.6.11")
-                    implementation("org.jetbrains.compose.desktop:desktop-jvm-macos-arm64:1.6.11")
-                    implementation("org.jetbrains.compose.desktop:desktop-jvm-linux-x64:1.6.11")
-                    implementation("org.jetbrains.compose.desktop:desktop-jvm-windows-x64:1.6.11")
-                }
-
-                testing {
-                    dependencies {
-                        implementation("org.jetbrains.compose.ui:ui-test-junit4:1.6.11")
-
-                        runtimeOnly("org.junit.vintage:junit-vintage-engine:5.12.0")
-                    }
+                    runtimeOnly("org.junit.vintage:junit-vintage-engine:5.12.0")
                 }
             }
         }
@@ -151,3 +149,4 @@ desktopComposeApp {
         }
     }
 }
+
